Finish README
This commit has the fully functioning code with all features, but with the depreciating mysql-connection type, and image capturing and so on. Future development will focus on changing to the new msqli-connection or PDO (PHP Data Object) which is better and faster. Reference page to learn more about PDO is http://code.tutsplus.com/tutorials/why-you-should-be-using-phps-pdo-for-database-access--net-12059